Add files via upload
Created a random dataset using the random and randint functions in numpy along with for loops. A counter is used to generate 100 random pieces of data for each variable (age, gender, county, year). Due to some of the variables not being measurable it was difficult to generate 100 random values as I couldn't find a function in numpy.random package.

while counter < 100:
    counterint     = 0
    results        = random.choice(gender)
    resultscounty  = random.choice(county)
    resultsyear    = random.choice(a)
    age            = randint(15, 75, 1)
    x.add_row([results, resultscounty, age, resultsyear])
    counter       += 1
    print(x)


the above code is used to generate a table format, however due to the counter the titles are inserted after each count +1, until 100 is reached, therefore it is not until the last one where the table looks ideal.